[REFACT] ProductAdditionalOption 제거
