+++ src/main/java/com/programmers/smrtstore/domain/order/orderSheet/presentation/dto/vo/OrderSheetProductOption.java
@@ -1,9 +1,9 @@
 package com.programmers.smrtstore.domain.order.orderSheet.presentation.dto.vo;
 
-import com.programmers.smrtstore.domain.product.domain.entity.ProductOption;
+import com.programmers.smrtstore.domain.product.domain.entity.ProductDetailOption;
 
 public class OrderSheetProductOption {
     private Long orderedProductOptionId;
-    private ProductOption productOption;
+    private ProductDetailOption productOption;
 
 }
