+++ src/main/java/com/programmers/smrtstore/domain/product/domain/entity/Product.java
@@ -83,9 +83,6 @@ public class Product {
     @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.EAGER)
     private final List<ProductDetailOption> productDetailOptions = new ArrayList<>();
 
-    @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, orphanRemoval = true)
-    private final List<ProductAdditionalOption> productAdditionalOptions = new ArrayList<>();
-
     @Column(name = "combination_yn", nullable = false)
     @JdbcTypeCode(SqlTypes.TINYINT)
     private boolean combinationYn = false;
@@ -140,10 +137,6 @@ public class Product {
         productDetailOptions.add(detailOption);
     }
 
-    public void addAdditionalOption(ProductAdditionalOption additionalOption) {
-        productAdditionalOptions.add(additionalOption);
-    }
-
     public void increaseStockQuantity(Integer quantity) {
         if (this.combinationYn) {
             throw new ProductException(ErrorCode.PRODUCT_USE_OPTION);
@@ -161,12 +154,6 @@ public class Product {
             .addStockQuantity(quantity);
     }
 
-    public void increaseAdditionalStockQuantity(Integer quantity, Long productOptionId) {
-        productAdditionalOptions.stream().filter(option -> option.getId().equals(productOptionId))
-            .findFirst()
-            .orElseThrow(() -> new ProductException(ErrorCode.PRODUCT_OPTION_NOT_FOUND))
-            .addStockQuantity(quantity);
-    }
 
     public void decreaseStockQuantity(Integer quantity) {
         if (this.combinationYn) {
@@ -187,14 +174,7 @@ public class Product {
             .removeStockQuantity(quantity);
     }
 
-    public void decreaseAdditionalStockQuantity(Integer quantity, Long productOptionId) {
-        productAdditionalOptions.stream().filter(option -> option.getId().equals(productOptionId))
-            .findFirst()
-            .orElseThrow(() -> new ProductException(ErrorCode.PRODUCT_OPTION_NOT_FOUND))
-            .removeStockQuantity(quantity);
-    }
-
-    public void removeOption(Long productOptionId) {
+    public void removeDetailOption(Long productOptionId) {
         if (!this.combinationYn) {
             throw new ProductException(ErrorCode.PRODUCT_USE_SINGLE_OPTION);
         }
@@ -204,17 +184,10 @@ public class Product {
         productDetailOptions.remove(productOption);
     }
 
-    public void removeAdditionalOption(Long productOptionId) {
-        var additionalOption = productAdditionalOptions.stream()
-            .filter(option -> option.getId().equals(productOptionId)).findAny()
-            .orElseThrow(() -> new ProductException(ErrorCode.PRODUCT_OPTION_NOT_FOUND));
-        productAdditionalOptions.remove(additionalOption);
-    }
-
     public Integer getStockQuantity() {
         if (combinationYn) {
             return productDetailOptions.stream()
-                .map(ProductOption::getStockQuantity)
+                .map(ProductDetailOption::getStockQuantity)
                 .reduce(0, Integer::sum);
         }
         return productDetailOptions.stream().findAny()
