+++ src/main/java/com/programmers/smrtstore/domain/product/domain/entity/enums/OptionType.java
@@ -2,8 +2,6 @@ package com.programmers.smrtstore.domain.product.domain.entity.enums;
 
 public enum OptionType {
     SINGLE,         // 단일 옵션 <- 옵션 없는 상품의 경우
-    COMBINATION,     // 조합 옵션
-
-    ADDITIONAL
+    COMBINATION     // 조합 옵션
 }
