+++ src/main/java/com/programmers/smrtstore/domain/product/presentation/ProductController.java
@@ -1,19 +1,14 @@
 package com.programmers.smrtstore.domain.product.presentation;
 
-import com.programmers.smrtstore.domain.product.application.ProductAdditionalService;
 import com.programmers.smrtstore.domain.product.application.ProductDetailService;
 import com.programmers.smrtstore.domain.product.application.ProductService;
-import com.programmers.smrtstore.domain.product.application.dto.req.ProductAdditionalOptionRequest;
 import com.programmers.smrtstore.domain.product.application.dto.req.ProductDetailOptionRequest;
 import com.programmers.smrtstore.domain.product.presentation.dto.req.CreateProductAPIRequest;
-import com.programmers.smrtstore.domain.product.presentation.dto.req.CreateProductAdditionalOptionAPIRequest;
 import com.programmers.smrtstore.domain.product.presentation.dto.req.CreateProductDetailOptionAPIRequest;
 import com.programmers.smrtstore.domain.product.presentation.dto.req.ProductDiscountRatioAPIRequest;
-import com.programmers.smrtstore.domain.product.presentation.dto.req.UpdateAdditionalOptionAPIRequest;
 import com.programmers.smrtstore.domain.product.presentation.dto.req.UpdateDetailOptionAPIRequest;
 import com.programmers.smrtstore.domain.product.presentation.dto.req.UpdateProductAPIRequest;
 import com.programmers.smrtstore.domain.product.presentation.dto.res.ProductAPIResponse;
-import com.programmers.smrtstore.domain.product.presentation.dto.res.ProductAdditionalOptionAPIResponse;
 import com.programmers.smrtstore.domain.product.presentation.dto.res.ProductDetailAPIResponse;
 import com.programmers.smrtstore.domain.product.presentation.dto.res.ProductDetailOptionAPIResponse;
 import com.programmers.smrtstore.domain.product.presentation.dto.res.ProductDiscountAPIResponse;
@@ -40,7 +35,6 @@ public class ProductController {
 
     private final ProductService productService;
     private final ProductDetailService productDetailService;
-    private final ProductAdditionalService productAdditionalService;
     private final ReviewService reviewService;
 
     @PostMapping
@@ -62,16 +56,6 @@ public class ProductController {
         var result = productDetailService.addProductDetailOption(productId, request.toRequest());
         return ResponseEntity.ok(ProductDetailOptionAPIResponse.from(result));
     }
-
-    @Secured("ROLE_ADMIN")
-    @PostMapping("/{productId}/additionalOptions")
-    public ResponseEntity<ProductAdditionalOptionAPIResponse> addNewAdditionalOption(
-        @PathVariable Long productId,
-        @Valid @RequestBody CreateProductAdditionalOptionAPIRequest request) {
-        var result = productAdditionalService.addAdditionalOption(productId, request.toRequest());
-        return ResponseEntity.ok(ProductAdditionalOptionAPIResponse.from(result));
-    }
-
     // ProductDetailPage
 
     @GetMapping
@@ -122,35 +106,10 @@ public class ProductController {
     }
 
     @Secured("ROLE_ADMIN")
-    @PutMapping("/{productId}/additionalOptions/{optionId}")
-    public ResponseEntity<ProductAdditionalOptionAPIResponse> productAdditionalOptionUpdate(
-        @PathVariable Long productId, @PathVariable Long optionId,
-        @Valid @RequestBody UpdateAdditionalOptionAPIRequest request) {
-        var result = productAdditionalService.updateAdditionalOption(
-            ProductAdditionalOptionRequest.builder()
-                .productId(productId)
-                .optionId(optionId)
-                .quantity(request.getQuantity())
-                .price(request.getPrice())
-                .groupName(request.getGroupName())
-                .name(request.getName())
-                .build());
-        return ResponseEntity.ok(ProductAdditionalOptionAPIResponse.from(result));
-    }
-
-    @Secured("ROLE_ADMIN")
     @DeleteMapping("/{productId}/detailOptions/{optionId}")
     public ResponseEntity<Long> removeDetailOption(@PathVariable Long productId,
         @PathVariable Long optionId) {
         var result = productDetailService.removeProductOption(productId, optionId);
-        return ResponseEntity.ok(result);
-    }
-
-    @Secured("ROLE_ADMIN")
-    @DeleteMapping("/{productId}/additionalOptions/{optionId}")
-    public ResponseEntity<Long> removeAdditionalOption(@PathVariable Long productId,
-        @PathVariable Long optionId) {
-        var result = productAdditionalService.removeAdditionalOption(productId, optionId);
         return ResponseEntity.ok(result);
     }
